cleaned examples and pom.xml files , no added features
