+++ .gitignore
@@ -34,80 +34,8 @@
 /app-layout-addon/src/main/resources/META-INF/resources/frontend/com/github/appreciated/app-layout/app-layout.css.map
 app-layout-addon/src/main/resources/META-INF/resources/frontend/com/github/appreciated/app-layout/app-layout.css.map
 app-layout-examples/app-layout-plain-example/target/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/cache/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/classes/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/frontend/bower_components/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/frontend/com/github/appreciated/app-layout/app-layout.css.map
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/frontend/com/github/appreciated/app-layout/left/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/frontend/com/github/appreciated/app-layout/styles/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/frontend/com/github/appreciated/app-layout/top/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/frontend/datepickerConnector.js
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/frontend/flow-component-renderer.html
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/frontend/gridConnector.js
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/frontend/images/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/frontend/ironListConnector.js
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/frontend/textConnector.js
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/frontend/vaadin-grid-flow-selection-column.html
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/generated-sources/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/gulpfile.js
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/maven-status/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/.bin/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/.yarn-integrity
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/@types/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/acorn-jsx/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/acorn/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/ansi-align/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/ansi-cyan/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/ansi-escape-sequences/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/ansi-gray/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/ansi-red/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/ansi-regex/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/ansi-styles/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/ansi-wrap/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/any-promise/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/archy/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/arr-diff/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/arr-flatten/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/arr-union/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/array-back/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/array-differ/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/array-each/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/array-find-index/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/array-slice/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/array-union/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/array-uniq/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/array-unique/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/assign-symbols/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/async/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/atob/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-code-frame/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-core/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-generator/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-helper-builder-binary-assignment-operator-visitor/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-helper-call-delegate/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-helper-define-map/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-helper-evaluate-path/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-helper-explode-assignable-expression/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-helper-flip-expressions/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-helper-function-name/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-helper-get-function-arity/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-helper-hoist-variables/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-helper-is-nodes-equiv/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-helper-is-void-0/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-helper-mark-eval-scopes/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-helper-optimise-call-expression/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-helper-regex/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-helper-remap-async-to-generator/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-helper-remove-or-void/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-helper-replace-supers/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-helper-to-multiple-sequence-expressions/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-helpers/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-messages/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-plugin-check-es2015-constants/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-plugin-external-helpers/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-plugin-minify-builtins/
-app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-plugin-minify-constant-folding/
+app-layout-ui-tests/app-layout-ui-test-left-responsive/target/
+
 app-layout-ui-tests/app-layout-ui-tes
 t-left-responsive/target/node_modules/babel-plugin-minify-dead-code-elimination/
 app-layout-ui-tests/app-layout-ui-test-left-responsive/target/node_modules/babel-plugin-minify-flip-comparisons/
