+++ app-layout-addon/src/main/java/com/github/appreciated/app/layout/builder/interfaces/ComponentFactory.java
@@ -3,7 +3,9 @@ package com.github.appreciated.app.layout.builder.interfaces;
 import com.vaadin.flow.component.HasElement;
 
 /**
- * An interface for classes that are able to give a specific output instance of the type {@link T} for an instance of the type {@link V} as input that is required to extend {@link Component}
+ * An interface for classes that are able to give a specific output instance
+ * of the type {@link T} for an instance of the
+ * type {@link V} as input that is required to extend {@link Component}
  *
  * @param <T> the output
  * @param <V> the input
