+++ app-layout-addon/src/main/java/com/github/appreciated/app/layout/component/appmenu/left/builder/LeftAppMenuBuilder.java
@@ -70,7 +70,7 @@ public class LeftAppMenuBuilder {
 
     public NavigationElementContainer build() {
         LeftMenuComponent menu = new LeftMenuComponent();
-        menu.add(components.toArray(new Component[components.size()]));
+        menu.add(components.toArray(new Component[0]));
         return menu;
     }
 }
