+++ app-layout-addon/src/main/java/com/github/appreciated/app/layout/notification/component/NotificationView.java
@@ -61,7 +61,7 @@ public class NotificationView extends PaperCard {
         add(headerLine);
         add(descriptionWrapper);
         if (clickEvent != null) {
-            getElement().addEventListener("click", domEvent -> clickEvent.handleEvent(domEvent));
+            getElement().addEventListener("click", clickEvent::handleEvent);
         }
         getElement().getClassList().add(info.getStyle());
         if (clickEvent != null) {
