+++ app-layout-examples/app-layout-annotation-example/src/main/java/com/github/appreciated/example/AbstractView.java
@@ -4,18 +4,22 @@ import com.vaadin.flow.component.html.Label;
 import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
 
 abstract class AbstractView extends HorizontalLayout {
-    public AbstractView() {
-        HorizontalLayout layout = new HorizontalLayout();
-        layout.setSizeFull();
-        layout.setMargin(false);
-        setMargin(false);
-        Label label = new Label("< " + getViewName() + " >");
-        layout.add(label);
-        layout.setAlignItems(Alignment.CENTER);
-        add(layout);
-        setSizeFull();
-        getElement().getStyle().set("overflow", "auto");
-    }
 
-    abstract String getViewName();
+  public AbstractView() {
+
+    HorizontalLayout layout = new HorizontalLayout();
+    layout.setSizeFull();
+    layout.setMargin(false);
+    layout.add(new Label("< " + getViewName() + " >"));
+    layout.setAlignItems(Alignment.CENTER);
+
+    add(layout);
+    setMargin(false);
+    setSizeFull();
+    getElement()
+        .getStyle()
+        .set("overflow", "auto");
+  }
+
+  abstract String getViewName();
 }
