+++ app-layout-examples/app-layout-annotation-example/src/main/java/com/github/appreciated/example/SubContent.java
@@ -9,7 +9,7 @@ import com.vaadin.flow.router.Route;
  * If you choose the route properly the addon supports back navigation. Also if the Subcontent does not appear in the Menu, the parent route will be highlighted.
  */
 
-@Route(value = "view2/subcontent", layout = AppLayoutWrapperView.class)
+@Route(value = "view2/subcontent", layout = MainAppLayout.class)
 public class SubContent extends VerticalLayout {
 
     public SubContent() {
