+++ app-layout-examples/app-layout-annotation-example/src/main/java/com/github/appreciated/example/View3.java
@@ -5,7 +5,7 @@ import com.github.appreciated.app.layout.annotations.Icon;
 import com.vaadin.flow.component.icon.VaadinIcon;
 import com.vaadin.flow.router.Route;
 
-@Route(value = "view3", layout = MainView.class)
+@Route(value = "view3", layout = MainAppLayout.class)
 @Caption("Contact")
 @Icon(VaadinIcon.CONNECT)
 public class View3 extends AbstractView {
