+++ app-layout-examples/app-layout-annotation-example/src/main/java/com/github/appreciated/example/View4.java
@@ -5,7 +5,7 @@ import com.github.appreciated.app.layout.annotations.Icon;
 import com.vaadin.flow.component.icon.VaadinIcon;
 import com.vaadin.flow.router.Route;
 
-@Route(value = "view4", layout = MainView.class)
+@Route(value = "view4", layout = MainAppLayout.class)
 @Caption("More")
 @Icon(VaadinIcon.COG)
 public class View4 extends AbstractView {
