+++ app-layout-examples/app-layout-annotation-example/src/main/java/com/github/appreciated/example/View5.java
@@ -5,7 +5,7 @@ import com.github.appreciated.app.layout.annotations.Icon;
 import com.vaadin.flow.component.icon.VaadinIcon;
 import com.vaadin.flow.router.Route;
 
-@Route(value = "view5", layout = MainView.class)
+@Route(value = "view5", layout = MainAppLayout.class)
 @Caption("Menu")
 @Icon(VaadinIcon.MENU)
 public class View5 extends AbstractView {
