+++ app-layout-examples/app-layout-plain-example/src/main/java/com/github/appreciated/example/View3.java
@@ -2,7 +2,7 @@ package com.github.appreciated.example;
 
 import com.vaadin.flow.router.Route;
 
-@Route(value = "view3", layout = AppLayoutWrapperView.class)
+@Route(value = "view3", layout = MainAppLayout.class)
 public class View3 extends AbstractView {
     @Override
     String getViewName() {
