+++ app-layout-examples/app-layout-plain-example/src/main/java/com/github/appreciated/example/View4.java
@@ -2,7 +2,7 @@ package com.github.appreciated.example;
 
 import com.vaadin.flow.router.Route;
 
-@Route(value = "view4", layout = AppLayoutWrapperView.class)
+@Route(value = "view4", layout = MainAppLayout.class)
 public class View4 extends AbstractView {
     @Override
     String getViewName() {
